Test display limit with thread limit
Prototype of running a limited number of threads, either manually or to
avoid mashing stdout / scrollback, each with its own queue.  This is
perhaps semi-synchronous rather than completely asynchronous, but it's a
start -- possibly will need to work out how to schedule stdout lines if
this index-based restriction is not applied.

* Split tasks across multiple queues
* Limit threads by stdout lines or by manual restriction